# dist/api.yml
openapi: 3.0.0
info:
  title: api
  description: API for  project
  version: 1.0.0
...
paths:
...
  /users/favoritePlayers:
    get:
      tags:
        - Users
      summary: Return the favorite players that were saved by user
      parameters:
      responses:
        "200":
          description: data about the user's favorites players returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerPreview"
        "401":
          description: user is not authorized
    post:
      summary: save player in the favorites list of a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                playerId:
                  type: string
                  description: player's id which the user saved
                  example: 85
      responses:
        "201":
          description: The player successfully saved as favorite
        "401":
          description: user is not authorized
...
  /users/favoriteGames:
    get:
      tags:
        - Users
      summary: Return the favorite games of the user
      responses:
        "200":
          description: returned data about the user's favorites games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
        "401":
          description: user is not authorized

    post:
      summary: save game in the favorites of a user
      tags:
        - Users
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                gameID:
                  type: string
                  description: game ID the user want to save
                  example: 7
      responses:
        "201":
          description: The game was successfully added to the favorite list
        "401":
          description: user is not authorized
  /users/lastSearch:
    get:
      tags:
        - Users
      summary: Return user's last search
      responses:
        "200":
          description: user's last search return successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/lastSearch"
        "204":
          description: No search was found
        
        "401":
          description: user is not authorized
...
components:
  schemas:
...
    PlayerPreview:
      type: object
      properties:
        name:
          type: string
          description: player's name
          example: Sten Michael Grytebust
        team_name:
          type: string
          description: player's team name
          example: København
        image:
          type: string
          description: player's image
          format: URL
          example: https://cdn.sportmonks.com/images/soccer/players/11/84587.png
        position:
          type: integer
          description: player's position in team
          example: 1
...
    Game:
      type: object
      properties:
        local_team:
          type: string
          description: local team name
          example: Midtjylland
        visitor_team:
          type: string
          description: visitor team name
          example: København
        local_team_score:
          type: integer
          description: local team score
          example: 0
        visitor_team_score: 
          type: integer
          description: visitor team score
          example: 1
        game_date:
          type: string
          description: date of the game
          example: '2019-09-10 11:30:00'
        field_name:
          type: string
          description: field name of the game
          example: Bloomfield
        referee_name:
          type: string
          description: referee name of the game
          example: Tom 
        event:
          type: array
          items:
            $ref: "#/components/schemas/Event"
...
    eventType: 
      type: string 
      enum:
        - goal
        - injury
        - offense
        - offside
        - redcard
        - substitution
        - yellowcard
    Event:
      type: object
      properties:
        event_time:
          type: string
          description: date and time of game
          example: '2017-08-10 11:30:00'
        
        minute_in_game:
          type: int
          description: The minute where the event accured
          example: 15
          
        event_description:
          type: string 
          description: Description of the event
          example: messi scores a goal
          
        type:
          $ref: '#/components/schemas/eventType'
          description: type of event 

        game_id:
          type: int
          description: id of the game the event belongs to
          example: 10
    lastSearch:
      type: object
      properties:
        query:
          type: string
          description: last query the user search 
          example: /James